Reorganize some tests into separate module
Use SimpleTestCase to reduce the overhead of using a database
transaction.
